Changed the model datatype of the income fields to BigDecimal. Removed the comma thousands separator from the format applied to the incomes in the fields.

+++ app/models/EligibilityCalculatorInput.scala
@@ -16,4 +16,9 @@
 
 package models
 
-case class EligibilityCalculatorInput(country: String, transferorIncome: Int, recipientIncome: Int)
+case class EligibilityCalculatorInput(
+                                       country: String,
+                                       transferorIncome: BigDecimal,
+                                       recipientIncome: BigDecimal
+                                     )
+
